requirements-dev.txt
# SURF Browser Service - Development Dependencies

# Include all core dependencies
-r requirements.txt

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-httpx>=0.26.0

# Code Quality
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.1

# Optional: Enhanced monitoring
prometheus-client>=0.19.0

